Changing the State column of the Table
Adding the file links to the State column and renaming the State
column to Results

XXX: The input dataset steps do not have job ids
There needs to be an umabmiguous way of associating a input data
set step with the correct file that was uploaded to that step.

XXX: Whenever the galaxy server is restarted the workflows need
to be repulled from galaxy. This shouldn't have to happen.